+++ src/app/cart/cart.module.ts
@@ -3,8 +3,9 @@ import { CommonModule } from '@angular/common';
 import { CartListComponent } from './components/cart-list/cart-list.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { CartService } from './services/cart.service';
-import { ProductsModule } from '../products/products.module';
 import { MaterialModule } from '../material/material.module';
+import { FormsModule } from '@angular/forms';
+import { SharedModule } from '../shared/shared.module';
 
 
 const components = [
@@ -18,7 +19,9 @@ const providers = [
   declarations: components,
   imports: [
     CommonModule,
+    FormsModule,
     MaterialModule,
+    SharedModule,
   ],
   providers,
   exports: components