+++ src/app/shared/shared.module.ts
@@ -1,14 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CategoryPipe } from './pipes/category.pipe';
+import { HighlightDirective } from './directives/highlight.directive';
 
 
 
 @NgModule({
-  declarations: [CategoryPipe],
+  declarations: [CategoryPipe, HighlightDirective],
   imports: [
     CommonModule,
   ],
-  exports: [CategoryPipe]
+  exports: [CategoryPipe, HighlightDirective]
 })
 export class SharedModule { }